build.js with webpack 2.0
