+++ apps/wmb/priv/www/static/.eslintrc.js
@@ -1,0 +1,7 @@
+module.exports = {
+    "extends": "standard",
+    "plugins": [
+        "standard",
+        "promise"
+    ]
+};
