+++ apps/wmb/priv/www/static/js/src/constants.js
@@ -1,0 +1,3 @@
+export const pathToAlbumsApi = '/api/albums/';
+export const pathToRandomApi = '/api/random/';
+export const pathToTracksApi = '/api/tracks/';
