+++ apps/wmb/priv/www/static/package.json
@@ -9,7 +9,7 @@
     "babel-eslint": "^5.0.0-beta6",
     "babel-loader": "^6.2.1",
     "babel-preset-es2015": "^6.3.13",
-    "babel-preset-react": "^6.3.13",
+    "babel-preset-stage-0": "^6.22.0",
     "css-loader": "^0.21.0",
     "es5-shim": "^4.4.0",
     "es6-promise": "^3.0.2",
@@ -25,9 +25,37 @@
     "webpack": "^2.2.1",
     "whatwg-fetch": "^0.10.1"
   },
-  "devDependencies": {},
+  "devDependencies": {
+    "better-npm-run": "0.0.14",
+    "eslint": "^1.10.3",
+    "eslint-config-standard": "^6.2.1",
+    "eslint-plugin-promise": "^3.4.2",
+    "eslint-plugin-standard": "^2.0.1"
+  },
   "scripts": {
-    "test": "node js/test/index.spec.js"
+    "build:dist": "better-npm-run build:dist",
+    "build:prod": "better-npm-run build:prod",
+    "test": "better-npm-run test"
+  },
+  "betterScripts": {
+    "build:dist": {
+      "command": "webpack --config webpack.config.dev.js --progress --colors",
+      "env": {
+        "NODE_ENV": "development"
+      }
+    },
+    "build:prod": {
+      "command": "webpack --config webpack.config.prod.js --progress --colors",
+      "env": {
+        "NODE_ENV": "production"
+      }
+    },
+    "test": {
+      "command": "karma start",
+      "env": {
+        "NODE_ENV": "test"
+      }
+    }
   },
   "author": "mcurrr",
   "license": "ISC"
